pyproject.toml
[tool.poetry]
name = "plateforme-data-demo-python"
version = "0.1.0"
description = ""
authors = ["sylvain tacquet <sylvain.tacquet@effidic.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy="1.25.0"
python-dotenv="1.0.0"
cryptography="38.0.4"
pre-commit = "^3.4.0"
pytest = "^7.4.2"
toml = "^0.10.2"
pylint = ">=2.17.5, <3.0.0"
flake8 = "6.1.0"
pip = "23.2.1"
black = "^23.3.0"
requests = "^2.31.0"
google-cloud = "^0.34.0"
google = "^3.0.0"
google-api-python-client = "^2.119.0"
google-cloud-storage = "^2.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
